Reorganize tests per review
