+++ packages/icon-button/package.json
@@ -7,10 +7,12 @@
   "homepage": "https://hig.autodesk.com",
   "main": "build/index.js",
   "module": "build/index.es.js",
+  "dependencies": {
+    "@hig/icon": "^0.1.0-alpha"
+  },
   "devDependencies": {
     "@hig/scripts": "^0.1.0",
-    "@hig/styles": "^0.1.0-alpha",
-    "@hig/icon": "^0.1.0-alpha"
+    "@hig/styles": "^0.1.0-alpha"
   },
   "peerDependencies": {
     "react": "^15.4.1",
