+++ packages/icon-button/src/index.test.js
@@ -1,0 +1,7 @@
+import { default as iconButtonTypes } from "./types";
+
+describe("IconButton/index", () => {
+  it("has type constants", () => {
+    expect(iconButtonTypes).toBeDefined();
+  });
+});
