--- packages/icons/src/icons.test.js
@@ -1,13 +1,0 @@
-import * as icons from "./index";
-
-describe("icons", () => {
-  describe("default", () => {
-    it("is a hash table of SVG markup", () => {
-      const ICON_MATCHER = /^<svg/;
-
-      Object.values(icons.default).forEach(icon => {
-        expect(icon).toMatch(ICON_MATCHER);
-      });
-    });
-  });
-});
